+++ internal/util.go
@@ -19,3 +19,29 @@ func SliceNextElem(v reflect.Value) reflect.Value {
 	v.Set(reflect.Append(v, reflect.Zero(elemType)))
 	return v.Index(v.Len() - 1)
 }
+
+func ToUpper(s string) string {
+	if isUpper(s) {
+		return s
+	}
+
+	b := make([]byte, len(s))
+	for i := range b {
+		c := s[i]
+		if c >= 'a' && c <= 'z' {
+			c -= 'a' - 'A'
+		}
+		b[i] = c
+	}
+	return string(b)
+}
+
+func isUpper(s string) bool {
+	for i := 0; i < len(s); i++ {
+		c := s[i]
+		if c >= 'a' && c <= 'z' {
+			return false
+		}
+	}
+	return true
+}