+++ types/append.go
@@ -6,8 +6,6 @@ import (
 	"reflect"
 	"strconv"
 	"time"
-
-	"gopkg.in/pg.v5/internal/parser"
 )
 
 func Append(b []byte, v interface{}, quote int) []byte {
@@ -178,57 +176,6 @@ func appendDriverValuer(b []byte, v driver.Valuer, quote int) []byte {
 		return AppendError(b, err)
 	}
 	return Append(b, value, quote)
-}
-
-func AppendField(b []byte, field string, quote int) []byte {
-	return appendField(b, parser.NewString(field), quote)
-}
-
-func AppendFieldBytes(b []byte, field []byte, quote int) []byte {
-	return appendField(b, parser.New(field), quote)
-}
-
-func appendField(b []byte, p *parser.Parser, quote int) []byte {
-	var quoted bool
-	for p.Valid() {
-		c := p.Read()
-
-		switch c {
-		case '*':
-			if !quoted {
-				b = append(b, '*')
-				continue
-			}
-		case '.':
-			if quoted && quote == 1 {
-				b = append(b, '"')
-				quoted = false
-			}
-			b = append(b, '.')
-			if p.Skip('*') {
-				b = append(b, '*')
-			} else if quote == 1 {
-				b = append(b, '"')
-				quoted = true
-			}
-			continue
-		}
-
-		if !quoted && quote == 1 {
-			b = append(b, '"')
-			quoted = true
-		}
-		if quote == 1 && c == '"' {
-			b = append(b, '"', '"')
-		} else {
-			b = append(b, c)
-		}
-
-	}
-	if quote == 1 && quoted {
-		b = append(b, '"')
-	}
-	return b
 }
 
 func appendAppender(b []byte, v ValueAppender, quote int) []byte {